requirements.txt
Django==2.2.12
pytz==2019.3
model-bakery==1.1.0
tox==3.14.6
pytest==5.4.1
pytest-django==3.9.0
black==19.10b0
flake8==3.7.9
isort==4.3.21
mypy==0.770
psycopg2-binary==2.8.5
django-storages==1.9.1
freezegun == 0.3.15
djangorestframework==3.11.0
